Added timestamp experiments and fixed set_param to be able to receive array values for parameters

+++ src/tuner.h
@@ -8,7 +8,7 @@
 struct metrics {
 	float latency;
 	float throughput;
-	float fct;
+	long fct;
 };
 
 /* Parameter specific functions */
